the basics to my final quest.
i was following along with Mr. Cozort's runthrough of Shmup but after awhile i got a bit lost so i decided to make a new file in which i would try to use the shmup file as reference and use the kids can code videos to guide me along but im getting stuck with getting the code to work when i hit start.